backend/requirements.txt
flask
flask-cors
torch
transformers
peft
sentence-transformers
pinecone-client
python-dotenv
pdfplumber
